+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/book/domain/BookRegister.java
@@ -1,19 +1,26 @@
-package sigma.chackcheck.domain.bookBorrow;
+package sigma.chackcheck.domain.book.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import sigma.chackcheck.domain.user.User;
-
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class BookReserve {
+public class BookRegister {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+    private String title;
+    private String author;
+    private String publishYear;
+    // 출판사
+    private String publisher;
+    // 도서 이미지
+    private String imageURL;
+    // 도서 관련 정보
+    private String information;
 
     // 유저 다대일
 //    @ManyToOne
@@ -21,8 +28,4 @@ public class BookReserve {
 //    private User user;
     @Column(name = "user_id")
     private Long userId;
-    // 도서 대출/반납 일대일
-    @OneToOne
-    @JoinColumn(name = "bookBorrow_id")
-    private BookBorrow bookBorrow;
 }
