+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/book/repository/BookDetailRepository.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.domain.book;
+package sigma.chackcheck.domain.book.repository;
 
 import org.springframework.stereotype.Repository;
