+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/book/service/BookService.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.service.book;
+package sigma.chackcheck.domain.book.service;
 
 import org.springframework.stereotype.Service;
