+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/bookBorrow/controller/BookBorrowController.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.controller;
+package sigma.chackcheck.domain.bookBorrow.controller;
 
 import org.springframework.stereotype.Controller;
