+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/bookBorrow/domain/BookReserve.java
@@ -1,27 +1,18 @@
-package sigma.chackcheck.domain.book;
+package sigma.chackcheck.domain.bookBorrow.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import sigma.chackcheck.domain.user.User;
+
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
-public class BookRegister {
+public class BookReserve {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String title;
-    private String author;
-    private String publishYear;
-    // 출판사
-    private String publisher;
-    // 도서 이미지
-    private String imageURL;
-    // 도서 관련 정보
-    private String information;
 
     // 유저 다대일
 //    @ManyToOne
@@ -29,4 +20,8 @@ public class BookRegister {
 //    private User user;
     @Column(name = "user_id")
     private Long userId;
+    // 도서 대출/반납 일대일
+    @OneToOne
+    @JoinColumn(name = "bookBorrow_id")
+    private BookBorrow bookBorrow;
 }
