--- BE/chackcheck/src/main/java/sigma/chackcheck/domain/user/Role.java
@@ -1,6 +1,0 @@
-package sigma.chackcheck.domain.user;
-
-public enum Role {
-    USER, ADMIN
-}
-
