+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/user/controller/UserController.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.controller;
+package sigma.chackcheck.domain.user.controller;
 
 import org.springframework.stereotype.Controller;
