+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/user/domain/Role.java
@@ -1,0 +1,6 @@
+package sigma.chackcheck.domain.user.domain;
+
+public enum Role {
+    USER, ADMIN
+}
+
