+++ BE/chackcheck/src/main/java/sigma/chackcheck/domain/user/service/UserService.java
@@ -1,4 +1,4 @@
-package sigma.chackcheck.service.user;
+package sigma.chackcheck.domain.user.service;
 
 import org.springframework.stereotype.Service;
